+++ view/frontend/requirejs-config.js
@@ -2,7 +2,6 @@ var config = {
   paths: {
     "paystand": "https://checkout.paystand.com/v4/js/paystand.checkout.js?env=live",
     "paystand-sandbox": "https://checkout.paystand.co/v4/js/paystand.checkout.js?env=sandbox",
-    //"paystand-sandbox": "https://localhost:3003/js/paystand.checkout.js?env=local_dev",
-    "paystand-checkout": "./PayStand_PayStandMagento/js/view/checkout/checkout"
+    //"paystand-sandbox": "https://checkout.paystand.co/v4/js/paystand.checkout.js?env=local_dev",
   }
 };
